+++ previous-versions/banking/pkg/models/operations/listbanktransactions.go
@@ -70,8 +70,6 @@ func (o *ListBankTransactionsRequest) GetQuery() *string {
 type ListBankTransactionsResponse struct {
 	// HTTP response content type for this operation
 	ContentType string
-	// Your `query` parameter was not correctly formed
-	ErrorMessage *shared.ErrorMessage
 	// HTTP response status code for this operation
 	StatusCode int
 	// Raw HTTP response; suitable for custom response parsing
@@ -85,13 +83,6 @@ func (o *ListBankTransactionsResponse) GetContentType() string {
 		return ""
 	}
 	return o.ContentType
-}
-
-func (o *ListBankTransactionsResponse) GetErrorMessage() *shared.ErrorMessage {
-	if o == nil {
-		return nil
-	}
-	return o.ErrorMessage
 }
 
 func (o *ListBankTransactionsResponse) GetStatusCode() int {